+++ lib/kiba-common/transforms/source_transform_adapter.rb
@@ -1,0 +1,8 @@
+class SourceTransformAdapter
+  def process(args)
+    args.shift.new(*args).each do |row|
+      yield(row)
+    end
+    nil
+  end
+end