+++ test/helper.rb
@@ -1,5 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/pride'
+require 'minitest/focus'
 
 require 'kiba-common/sources/enumerable'
 require_relative 'support/assert_called'